pyproject.toml
[project]
name = "f4-plugin"
version = "0.1.0"
description = "Flexion Fast Fail Filtering - RFP flag detection service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.92.0",
    "boto3>=1.42.86",
    "chromadb>=1.0",
    "gradio>=5.0",
    "pdfplumber>=0.11",
    "python-docx>=1.1",
    "sentence-transformers>=3.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src --cov-report=term-missing"

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "reportlab>=4.0",
    "ruff>=0.15.9",
]
training = [
    "torch>=2.0",
    "transformers>=4.45",
    "peft>=0.13",
    "trl>=0.12",
    "datasets>=3.0",
    "accelerate>=1.0",
]
